add main.js for initWallet
